Đã fix lỗi đăng kí thành công
