+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,9 @@
         android:theme="@style/Theme.SieuThiSach"
         tools:targetApi="31">
         <activity
+            android:name=".DangKi"
+            android:exported="false" />
+        <activity
             android:name=".Model.DangKi"
             android:exported="false" />
         <activity
